requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
aioxmpp

aiohttp

ruamel.yaml
